Add option to sort paraphrasing datasets set by length
This reduces the need for padding tokens during training and evaluation.
Note that the each minibatch has examples of similar length,
but the order of minibatches is still random. With batch size of 2,
this has no effect on training speed; however,
with larger batch sizes, the benefits are significant.
For instance, --sort_by_length gives a ~2.5x speedup for batch size of 32.
Surprisingly, this has a positive effect in terms of accuracy/BLEU score
on BART,  GPT2-medium, GPT2-small and distil-GPT2, but no effect
on GPT2-large.